requirements.txt
datapizza-ai>=0.1.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
